feat: rename module externalDb.js to idInModels.js
